fix problem of intent in program and competition

+++ app/src/main/java/com/example/mobile_pfe/programActivity/ProgramActivity.java
@@ -33,7 +33,7 @@ public class ProgramActivity extends AppCompatActivity {
         setContentView(R.layout.activity_program);
 
         if (getIntent().hasExtra("programId")) {
-            int programId = getIntent().getIntExtra("programId", 1);
+            Long programId = getIntent().getLongExtra("programId", 1L);
             Toast.makeText(this, "Received program ID: " + programId, Toast.LENGTH_SHORT).show();
             // Use the program ID as needed
 
@@ -41,7 +41,7 @@ public class ProgramActivity extends AppCompatActivity {
             ProgramService service = RetrofitInstance.getRetrofitInstance().create(ProgramService.class);
 
             /*Call the method with parameter in the interface to get the employee data*/
-            Call<Program> call = service.getById(programId);
+            Call<Program> call = service.getById(Math.toIntExact(programId));
 
             /*Log the URL called*/
             Log.wtf("URL Called", call.request().url() + "");
